+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/compose/Carousel.kt
@@ -2,35 +2,22 @@ package ai.mcpdirect.studio.app.compose
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.rememberPagerState
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
+import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
+import mcpdirectstudioapp.composeapp.generated.resources.Res
+import mcpdirectstudioapp.composeapp.generated.resources.keyboard_arrow_left
+import mcpdirectstudioapp.composeapp.generated.resources.keyboard_arrow_right
 import org.jetbrains.compose.resources.DrawableResource
 import org.jetbrains.compose.resources.painterResource
 
@@ -42,6 +29,7 @@ data class CarouselSlide(
 
 @Composable
 fun Carousel(
+    title: String,
     slides: List<CarouselSlide>,
     modifier: Modifier,
     onComplete: () -> Unit = {}
@@ -50,10 +38,14 @@ fun Carousel(
     val scope = rememberCoroutineScope()
 
     Column(
-        modifier = modifier,
+        modifier = modifier.padding(vertical = 64.dp),
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.Center
     ) {
+        Text(
+            text = title,
+            style = MaterialTheme.typography.displayMedium,
+        )
         // Image Pager
         HorizontalPager(
             state = pagerState,
@@ -66,13 +58,6 @@ fun Carousel(
         PageIndicators(
             pageCount = slides.size,
             currentPage = pagerState.currentPage,
-            modifier = Modifier.padding(16.dp)
-        )
-
-        // Navigation Buttons
-        PagerNavigation(
-            currentPage = pagerState.currentPage,
-            totalPages = slides.size,
             onPrevious = {
                 scope.launch {
                     pagerState.animateScrollToPage(pagerState.currentPage - 1)
@@ -87,11 +72,32 @@ fun Carousel(
                     onComplete()
                 }
             },
-            onSkip = onComplete,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp, vertical = 24.dp)
+            modifier = Modifier.padding(16.dp)
         )
+
+        // Navigation Buttons
+//        PagerNavigation(
+//            currentPage = pagerState.currentPage,
+//            totalPages = slides.size,
+//            onPrevious = {
+//                scope.launch {
+//                    pagerState.animateScrollToPage(pagerState.currentPage - 1)
+//                }
+//            },
+//            onNext = {
+//                if (pagerState.currentPage < slides.size - 1) {
+//                    scope.launch {
+//                        pagerState.animateScrollToPage(pagerState.currentPage + 1)
+//                    }
+//                } else {
+//                    onComplete()
+//                }
+//            },
+//            onSkip = onComplete,
+//            modifier = Modifier
+//                .fillMaxWidth()
+//                .padding(horizontal = 16.dp, vertical = 24.dp)
+//        )
     }
 }
 
@@ -107,18 +113,18 @@ fun CarouselSlideItem(slide: CarouselSlide) {
         // Title
         Text(
             text = slide.title,
-            style = MaterialTheme.typography.displayMedium,
+            style = MaterialTheme.typography.displaySmall,
             modifier = Modifier.padding(vertical = 16.dp),
 
         )
         // Image
-        Card(Modifier.height(320.dp)){
+        OutlinedCard(Modifier.width(600.dp)){
             Image(
                 painter = painterResource(slide.imageResource),
                 contentDescription = slide.title,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(320.dp),
+//                modifier = Modifier
+//                    .fillMaxWidth()
+//                    .height(320.dp),
                 contentScale = ContentScale.Fit
             )
         }
@@ -136,12 +142,22 @@ fun CarouselSlideItem(slide: CarouselSlide) {
 fun PageIndicators(
     pageCount: Int,
     currentPage: Int,
+    onPrevious: () -> Unit,
+    onNext: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     Row(
         modifier = modifier,
+        verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.spacedBy(8.dp)
     ) {
+        IconButton(
+            colors = IconButtonDefaults.filledIconButtonColors(),
+            enabled = currentPage > 0,
+            onClick = onPrevious,
+        ){
+            Icon(painterResource(Res.drawable.keyboard_arrow_left),contentDescription = null)
+        }
         repeat(pageCount) { index ->
             Box(
                 modifier = Modifier
@@ -156,68 +172,75 @@ fun PageIndicators(
                     )
             )
         }
+        IconButton(
+            colors = IconButtonDefaults.filledIconButtonColors(),
+            enabled = currentPage < pageCount - 1,
+            onClick = onNext,
+        ){
+            Icon(painterResource(Res.drawable.keyboard_arrow_right),contentDescription = null)
+        }
     }
 }
 
-@Composable
-fun PagerNavigation(
-    currentPage: Int,
-    totalPages: Int,
-    onPrevious: () -> Unit,
-    onNext: () -> Unit,
-    onSkip: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier,
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // Previous Button (only show if not on first page)
-        if (currentPage > 0) {
-            Button(
-                onClick = onPrevious,
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = Color.Transparent,
-                    contentColor = MaterialTheme.colorScheme.primary
-                ),
-                elevation = null
-            ) {
-                Text("Previous")
-            }
-        }
-        Spacer(Modifier.weight(1f))
-//        else {
-//            // Skip Button (only on first page)
+//@Composable
+//fun PagerNavigation(
+//    currentPage: Int,
+//    totalPages: Int,
+//    onPrevious: () -> Unit,
+//    onNext: () -> Unit,
+//    onSkip: () -> Unit,
+//    modifier: Modifier = Modifier
+//) {
+//    Row(
+//        modifier = modifier,
+//        horizontalArrangement = Arrangement.SpaceBetween,
+//        verticalAlignment = Alignment.CenterVertically
+//    ) {
+//        // Previous Button (only show if not on first page)
+//        if (currentPage > 0) {
 //            Button(
-//                onClick = onSkip,
+//                onClick = onPrevious,
 //                colors = ButtonDefaults.buttonColors(
 //                    containerColor = Color.Transparent,
-//                    contentColor = MaterialTheme.colorScheme.onSurfaceVariant
+//                    contentColor = MaterialTheme.colorScheme.primary
 //                ),
 //                elevation = null
 //            ) {
-//                Text("Skip")
+//                Text("Previous")
 //            }
 //        }
-
-        // Next/Get Started Button
-
-//        Button(
+//        Spacer(Modifier.weight(1f))
+////        else {
+////            // Skip Button (only on first page)
+////            Button(
+////                onClick = onSkip,
+////                colors = ButtonDefaults.buttonColors(
+////                    containerColor = Color.Transparent,
+////                    contentColor = MaterialTheme.colorScheme.onSurfaceVariant
+////                ),
+////                elevation = null
+////            ) {
+////                Text("Skip")
+////            }
+////        }
+//
+//        // Next/Get Started Button
+//
+////        Button(
+////            onClick = onNext,
+////            modifier = Modifier.width(120.dp)
+////        ) {
+////            Text(
+////                text = if (currentPage < totalPages - 1) "Next" else "Get Started"
+////            )
+////        }
+//        if (currentPage < totalPages - 1) Button(
 //            onClick = onNext,
 //            modifier = Modifier.width(120.dp)
 //        ) {
 //            Text(
-//                text = if (currentPage < totalPages - 1) "Next" else "Get Started"
+//                text =  "Next"
 //            )
 //        }
-        if (currentPage < totalPages - 1) Button(
-            onClick = onNext,
-            modifier = Modifier.width(120.dp)
-        ) {
-            Text(
-                text =  "Next"
-            )
-        }
-    }
-}
+//    }
+//}